pyproject.toml
[tool.poetry]
name = "pandoc-glossary-filter"
version = "0.3.0"
description = "Pandoc filter used to generate a glossary and an acronym list common to a set of documents"
authors = ["Augusto Zanellato <augusto.zanellato@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-or-later"
packages = [{include = "pandoc_glossary_filter"}]

[tool.poetry.dependencies]
python = "^3.7.2"
pyyaml = "^6.0"
loguru = "^0.6.0"
panflute = "^2.2.3"

[tool.poetry.scripts]
pandoc-glossary = 'pandoc_glossary_filter:main'

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
pylint = "^2.15.5"

[tool.black]
line-length = 120
target-version = ['py37']

[tool.isort]
profile = "black"

[tool.pylint.main]
py-version = [3, 7]
disable = ["missing-module-docstring"]
[tool.pylint.format]
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
